✨ Feat: Api 함수 폴더링 작업
