+++ src/apis/core/errorHandler.js
@@ -1,0 +1,11 @@
+// 공용 에러 핸들러
+export const handleError = (error) => {
+  if (error.response) {
+    console.error('API Error:', error.response.status, error.response.data);
+  } else if (error.request) {
+    console.error('Network Error:', error.message);
+  } else {
+    console.error('Error:', error.message);
+  }
+  return Promise.reject(error);
+};
